moved retriever functionality to retriever.py
